Add files via upload
I built this script to aid an issue in my current job role. I handle a number of Network attached starage devices and it is imperative that correct ip addresses are placed in the host file, so that devices can communicate. This bash script defines a host path and also defines a specific header within the file. It then prompts a user to enter a valid ipv4 address, validates it, and removes any instance of an ip address in the file replacing it with the one entered by the user with sed.